requirements.txt
# Web框架
Flask==2.0.1

# WebSocket支持
Flask-SocketIO==5.3.6
eventlet==0.33.3

# HTTP请求库
requests==2.25.1

# 环境变量管理
python-dotenv==1.1.0

# 日志库
loguru==0.5.3